Fix compiler warnings across all modules
- Remove unused imports and parameters
- Add @Suppress annotations where warnings are intentional
- Fix deprecated API usages
- Remove redundant Elvis operator right-side expressions

+++ app/src/main/java/com/weatherwidget/ui/WeatherObservationsActivity.kt
@@ -537,7 +537,7 @@ class WeatherObservationsActivity : AppCompatActivity() {
             } else {
                 val displayTemp = if (useCelsius) com.weatherwidget.shared.util.TempUtils.fahrenheitToCelsius(item.temperature) else item.temperature
                 holder.temperature.text = String.format("%.1f°", displayTemp)
-                holder.temperature.setTextColor(obsTempToColor(item.temperature.toFloat()))
+                holder.temperature.setTextColor(obsTempToColor(item.temperature))
             }
             holder.condition.text = item.condition
         }
